+++ docs/guides/staging-servers.md
@@ -1,6 +1,6 @@
 ---
 tags:
-    - 👷 Staff
+  - 👷 Staff
 ---
 
 ## Data Manager Workflow
@@ -32,7 +32,7 @@ sequenceDiagram
     PR ->> SSB: Destroy server
 ```
 
-## Staging Sync Workflow
+## Chart Sync Workflow
 
 Once the work is merged, data manager should run `etl chart-sync` to migrate all charts to the live site. This command will sync all charts from staging to live as either draft charts or revisions.
